+++ app/html/comments.js
@@ -1,20 +1,12 @@
 const nest = require('depnest')
-const { h, Value, map, computed, when, resolve, throttle } = require('mutant')
+const { h, map, Struct, computed, throttle } = require('mutant')
 const get = require('lodash/get')
 
 exports.gives = nest('app.html.comments')
 
 exports.needs = nest({
-  'about.html.avatar': 'first',
-  'about.obs.name': 'first',
-  'backlinks.obs.for': 'first',
-  'feed.obs.thread': 'first',
+  'message.html.comment': 'first',
   'message.html.compose': 'first',
-  'message.html.markdown': 'first',
-  'message.html.timeago': 'first',
-  'message.html.likes': 'first',
-  'unread.sync.markRead': 'first',
-  'unread.sync.isUnread': 'first',
   'translations.sync.strings': 'first'
 })
 
@@ -28,11 +20,16 @@ exports.create = (api) => {
     // TODO - move this up into Patchcore
     const messagesTree = computed(throttle(messages, 200), msgs => {
       return msgs
-        .filter(msg => forkOf(msg) === undefined)
-        .map(threadMsg => {
-          const nestedReplies = msgs.filter(msg => forkOf(msg) === threadMsg.key)
-          threadMsg.replies = nestedReplies
-          return threadMsg
+        .filter(msg => forkOf(msg) === undefined) // exclude nested replies / forks
+        .filter(msg => msg.value.content.root) // exclude root message / blog
+        .map(comment => {
+          const nestedReplies = msgs.filter(msg => forkOf(msg) === comment.key)
+
+          return Struct({
+            comment: comment,
+            replies: nestedReplies
+          })
+          // return Object.assign({}, comment, { replies: nestedReplies })
         })
     })
 
@@ -55,92 +52,19 @@ exports.create = (api) => {
 
     return h('Comments', [
       // when(twoComposers, compose({ meta, shrink: true, canAttach: false })),
-      map(messagesTree, msg => Comment(msg, root, branch)),
+      map(
+        messagesTree,
+        msg => api.message.html.comment({ comment: msg.comment, replies: msg.replies, branch }),
+        {
+          comparer: (a, b) => {
+            if (a === undefined || b === undefined) return false
+
+            return a.comment() === b.comment() && a.replies().length === b.replies().length
+          }
+        }
+      ),
       compose({ meta, feedIdsInThread, shrink: false, canAttach: true, placeholder: strings.writeComment })
     ])
-  }
-
-  function Comment (msgObs, root, branch) {
-    const strings = api.translations.sync.strings()
-    const msg = resolve(msgObs)
-
-    const raw = get(msg, 'value.content.text')
-    var className = api.unread.sync.isUnread(msg) ? ' -unread' : ' -read'
-    api.unread.sync.markRead(msg)
-
-    if (!get(msg, 'value.content.root')) return
-
-    const { author, content } = msg.value
-
-    // // TODO - move this upstream into patchcore:feed.obs.thread ??
-    // // OR change strategy to use forks
-    // const backlinks = api.backlinks.obs.for(msg.key)
-    // const nestedReplies = computed(backlinks, backlinks => {
-    //   return backlinks.filter(backlinker => {
-    //     const { type, root } = backlinker.value.content
-    //     return type === 'post' && root === msg.key
-    //   })
-    // })
-
-    var nestedReplyCompose = Value(false)
-    const toggleCompose = () => nestedReplyCompose.set(!nestedReplyCompose())
-    const nestedReplyComposer = api.message.html.compose({
-      meta: {
-        type: 'post',
-        root,
-        fork: msg.key,
-        branch,
-        channel: content.channel
-      },
-      shrink: false,
-      canAttach: true,
-      canPreview: false,
-      placeholder: strings.writeComment
-    }, toggleCompose)
-
-    return h('Comment', { className }, [
-      h('div.left', api.about.html.avatar(author, 'tiny')),
-      h('div.right', [
-        h('section.context', [
-          h('div.name', api.about.obs.name(author)),
-          api.message.html.timeago(msg)
-        ]),
-        h('section.content', api.message.html.markdown(raw)),
-        when(msgObs.replies,
-          h('section.replies',
-            map(msgObs.replies, NestedComment)
-          )
-        ),
-        h('section.actions', [
-          h('div.reply', { 'ev-click': toggleCompose }, [
-            h('i.fa.fa-commenting-o')
-          ]),
-          api.message.html.likes(msg)
-        ]),
-        when(nestedReplyCompose, nestedReplyComposer)
-      ])
-    ])
-  }
-
-  function NestedComment (msgObs) {
-    const msg = resolve(msgObs)
-    const raw = get(msg, 'value.content.text')
-    if (!raw) return
-
-    const { author } = msg.value
-
-    return h('Comment -nested', [
-      h('div.left'),
-      h('div.right', [
-        h('section.context', [
-          h('div.name', api.about.obs.name(author)),
-          api.message.html.timeago(msg)
-        ]),
-        h('section.content', api.message.html.markdown(raw))
-      ])
-    ])
-
-    api.message.html.markdown(raw)
   }
 }
