+++ message/html/comment.mcss
@@ -1,0 +1,76 @@
+Comment {
+  display: flex
+
+  div.left {
+    margin-right: 1rem
+
+    div.Avatar {}
+  }
+
+  div.right {
+    flex-grow: 1
+
+    $borderBottomLight
+    padding-bottom: 1rem
+    margin-bottom: 1rem
+
+    section.context {
+      display: flex
+      align-items: baseline
+
+      div.name {
+        font-size: 1.2rem
+        margin-right: 1rem
+      }
+      div.Timeago {}
+    }
+
+    section.content {
+      font-size: .95rem
+      line-height: 1.4
+      div.Markdown {}
+    }
+
+    section.actions {
+      font-size: 1rem
+      color: #555
+      margin-right: .5rem
+
+      display: flex
+      justify-content: flex-end
+      align-items: baseline
+
+      div.reply {
+        cursor: pointer
+
+        margin-right: 1.5rem
+        i.fa {}
+      }
+
+      div.Likes { }
+    }
+
+    div.Compose {
+      margin: 1rem 0
+    }
+  }
+}
+
+Comment -nested {
+  padding: 1rem 1rem 0 1rem
+  $backgroundPrimaryText
+  $roundTop
+  $roundBottom
+
+  margin-bottom: 1rem
+
+  div.left { margin: 0 }
+  div.right {
+    padding: 0
+    border: 0
+    margin: 0
+
+  }
+}
+
+
