+++ package.json
@@ -41,7 +41,7 @@
     "medium-editor-markdown": "^2.6.0",
     "micro-css": "^2.0.1",
     "morphdom": "^2.3.3",
-    "mutant": "^3.21.2",
+    "mutant": "^3.22.1",
     "mutant-scroll": "0.0.5",
     "open-external": "^0.1.1",
     "patch-drafts": "0.0.6",
